Data Engineer/fact_table_keyword/requirements.txt
pandas
sqlalchemy
python-dotenv
os